refactor: move chart selectors
